# public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Fleet Finance Manger",
        "description": "Fleet Finance Manger - API for Android and IOS Development",
        "termsOfService": "All rights reserved to Greefi Technologies",
        "contact": {
            "email": "hello@greefitech.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:9000",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
        "/api/account/create": {
            "post": {
                "tags": [
                    "Account"
                ],
                "description": "Create New Account",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "account",
                        "in": "query",
                        "description": "Enter Account / Bank Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "HolderName",
                        "in": "query",
                        "description": "Enter Account holder name",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Invalid Data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/accounts": {
            "get": {
                "tags": [
                    "Account"
                ],
                "description": "Return all Accounts",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Invalid Data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/account/{id}/edit": {
            "get": {
                "tags": [
                    "Account"
                ],
                "description": "Return particular EAccount",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Invalid Data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/account/{id}/update": {
            "post": {
                "tags": [
                    "Account"
                ],
                "description": "Update Account",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "account",
                        "in": "query",
                        "description": "Enter Account / Bank Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "HolderName",
                        "in": "query",
                        "description": "Enter Account holder name",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Invalid Data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
}